requirements.txt
Pillow>=10.0.0
cryptography>=42.0.0
numpy>=1.26.0
customtkinter>=5.2.0
python-dotenv>=1.0.0
pyinstaller>=6.0.0
